requirements.txt
streamlit
mysql-connector-python
werkzeug
matplotlib
